Avoid failing postinstall if the target project has babel plugins. Closes #1
The problem was found to be the following:

Once reactuate is put into <target>/node_modules/reactuate, most of dependencies
are actually in <target>/node_modules. Therefore, when `babel-register` was
trying to load everything, it was not ignoring its own plugins as the path
to them did not contain node_modules, as it tries to use relative path, and
getting ../babel-plugin-* kind of path.

The current solution is not perfect. I manually converted postinstall script
(with babel-cli) to be a regular ES5 script and avoid the problem altogether.

Filed an issue at Babel: https://phabricator.babeljs.io/T7053